测试队列Task

+++ app/src/main/java/com/optima/plugin/host/thread/TestCallback.java
@@ -6,6 +6,6 @@ package com.optima.plugin.host.thread;
  */
 public interface TestCallback {
     void start();
-    void loading(int process,int total);
+    void loading(long process,long total);
     void finish();
 }
